SecureHoney/admin-panel/backend/requirements.txt
# Core FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1

# Redis for caching and sessions
redis[hiredis]==5.0.1

# Security and Authentication
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
cryptography==41.0.7
bcrypt==4.1.2

# Data validation and serialization
pydantic[email]==2.5.0

# HTTP client
httpx==0.25.2

# WebSocket support
websockets==12.0

# File handling
aiofiles==23.2.1

# Configuration and environment
python-dotenv==1.0.0
pyyaml==6.0.1

# Rate limiting
slowapi==0.1.9

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Email support
aiosmtplib==3.0.1

# Testing and development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0

# Advanced features
geoip2==4.7.0
cryptography==41.0.7
aiofiles==23.2.1
websockets==12.0

# Production server
gunicorn==21.2.0
